requirements.txt
playwright>=1.40.0
pillow>=10.0.0
openai>=1.0.0
click>=8.1.0
